+++ nnmnkwii/autograd/__init__.py
@@ -1,4 +1,5 @@
 from __future__ import with_statement, print_function, absolute_import
 
-from nnmnkwii.autograd._impl.mlpg import mlpg, MLPG
-from nnmnkwii.autograd._impl.modspec import modspec, ModSpec
+from ._impl.mlpg import mlpg, MLPG
+from ._impl.mlpg import unit_variance_mlpg, UnitVarianceMLPG
+from ._impl.modspec import modspec, ModSpec