+++ nnmnkwii/functions/__init__.py
@@ -1,5 +1,6 @@
 # coding: utf-8
 from __future__ import division, print_function, absolute_import
 
-from nnmnkwii.functions._impl.mlpg import mlpg, mlpg_grad
-from nnmnkwii.functions._impl.modspec import modspec, modphase
+from ._impl.mlpg import mlpg, mlpg_grad, build_win_mats, full_window_mat
+from ._impl.mlpg import unit_variance_mlpg_matrix, reshape_means
+from ._impl.modspec import modspec, modphase