+++ tests/test_autograd.py
@@ -1,14 +1,16 @@
 from __future__ import division, print_function, absolute_import
 
-from nnmnkwii.autograd._impl.mlpg import MLPG
+from nnmnkwii.autograd._impl.mlpg import MLPG, UnitVarianceMLPG
 from nnmnkwii.autograd._impl.modspec import ModSpec
+from nnmnkwii import functions as F
+from nnmnkwii import autograd as AF
 
 from torch.autograd import gradcheck
 from torch.autograd import Variable
+from torch import nn
 import torch
 import numpy as np
 from warnings import warn
-
 
 def _get_windows_set():
     windows_set = [
@@ -30,6 +32,69 @@ def _get_windows_set():
     ]
     return windows_set
 
+def test_functional_mlpg():
+    static_dim = 2
+    T = 10
+
+    for windows in _get_windows_set():
+        torch.manual_seed(1234)
+        means = torch.rand(T, static_dim * len(windows))
+        reshaped_means = torch.from_numpy(F.reshape_means(means.numpy(), static_dim))
+        variances = torch.ones(static_dim * len(windows))
+
+        y = F.mlpg(means.numpy(), variances.numpy(), windows)
+        y = Variable(torch.from_numpy(y), requires_grad=False)
+
+        means = Variable(means, requires_grad=True)
+        reshaped_means = Variable(reshaped_means, requires_grad=True)
+
+        y_hat = AF.mlpg(means, variances, windows)
+        assert np.allclose(y.data.numpy(), y_hat.data.numpy())
+
+        # Test backward pass
+        nn.MSELoss()(y_hat, y).backward()
+
+        R = torch.from_numpy(F.unit_variance_mlpg_matrix(windows, T))
+        y_hat = AF.unit_variance_mlpg(R, reshaped_means)
+        assert np.allclose(y.data.numpy(), y_hat.data.numpy())
+
+        nn.MSELoss()(y_hat, y).backward()
+
+def test_unit_variance_mlpg():
+    static_dim = 2
+    T = 10
+
+    for windows in _get_windows_set():
+        torch.manual_seed(1234)
+        # Meens, input for MLPG
+        means = Variable(torch.rand(T, static_dim * len(windows)),
+                         requires_grad=True)
+
+        # Input for UnitVarianceMLPG
+        # Equivalent:
+        # reshaped_means = means.view(
+        #    T, len(windows), -1).transpose(0, 1).contiguous().view(-1, static_dim)
+        reshaped_means = F.reshape_means(means.data.clone().numpy(), static_dim)
+        reshaped_means = Variable(torch.from_numpy(reshaped_means),
+                                  requires_grad=True)
+
+        # Compute MLPG matrix
+        R = F.unit_variance_mlpg_matrix(windows, T).astype(np.float32)
+        R = torch.from_numpy(R)
+
+        y = UnitVarianceMLPG(R)(reshaped_means)
+        # Unit variances
+        variances = torch.ones(static_dim * len(windows)
+                               ).expand(T, static_dim * len(windows))
+        y_hat = MLPG(variances, windows)(means)
+
+        # Make sure UnitVarianceMLPG and MLPG can get same result
+        # if we use unit variances
+        assert np.allclose(y.data.numpy(), y_hat.data.numpy())
+
+        inputs = (reshaped_means,)
+        assert gradcheck(UnitVarianceMLPG(R),
+                         inputs, eps=1e-3, atol=1e-3)
 
 def test_mlpg_gradcheck():
     # MLPG is performed dimention by dimention, so static_dim 1 is enough,
@@ -57,6 +122,20 @@ def test_mlpg_gradcheck():
         assert gradcheck(MLPG(variances, windows),
                          inputs, eps=1e-3, atol=1e-3)
 
+
+def test_mlpg_variance_expand():
+    static_dim = 2
+    T = 10
+
+    for windows in _get_windows_set():
+        torch.manual_seed(1234)
+        means = Variable(torch.rand(T, static_dim * len(windows)),
+                         requires_grad=True)
+        variances = torch.rand(static_dim * len(windows))
+        variances_expanded = variances.expand(T, static_dim * len(windows))
+        y = AF.mlpg(means, variances, windows)
+        y_hat = AF.mlpg(means, variances_expanded, windows)
+        assert np.allclose(y.data.numpy(), y_hat.data.numpy())
 
 def test_modspec_gradcheck():
     static_dim = 12