+++ NEWS.md
@@ -1,5 +1,9 @@
 # News for nnmnkwii
 
+## dev
+
+- Add new function `UnitVarianceMLPG` that can run on CPU/GPU. ([#21])
+
 ## v0.0.2 2017.08.18
 
 - hts io: Add support for full-context only label files
@@ -12,3 +16,4 @@
 - Initial release
 
 [#17]: https://github.com/r9y9/nnmnkwii/pull/17
+[#21]: https://github.com/r9y9/nnmnkwii/pull/21